LayoutTests/inspector/debugger-pause-in-eval-script.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script>

function doit()
{
    function callback(result)
    {
        if (result instanceof Array) {
            for (var i = 0; i < result.length; i++)
                output(result[i]);
        } else
            output(result);
        notifyDone();
    }
    evaluateInWebInspector("frontend_testPauseOnDebuggerStatement", callback);
}

function frontend_testPauseOnDebuggerStatement(testController)
{
    testController.waitUntilDone();

    WebInspector.showScriptsPanel();
    frontend_ensureDebuggerEnabled(debuggerWasEnabledCallback);

    var scriptToEvaluate = "function testFunction() {\n" +
                           "    debugger;\n" +
                           "}\n" +
                           "setTimeout(testFunction, 0);\n";

    function debuggerWasEnabledCallback()
    {
        testController.results.push("Debugger was enabled.");

        frontend_evaluateInConsole(scriptToEvaluate,
        function(result)
        {
            testController.results.push("Evaluated script in console.");
        });
    }

    frontend_addSniffer(WebInspector, "pausedScript", function(callFrames)
    {
        testController.results.push("Script execution paused.");
        frontend_checkExecutionLineWhenSourceFrameIsLoaded(testController, completeTest);
    });

    function completeTest()
    {
       if (WebInspector.panels.scripts.paused)
            WebInspector.panels.scripts._togglePause();
        testController.results.push("Resumed script execution.");
        testController.notifyDone();
    }
}

function frontend_ensureDebuggerEnabled(callback)
{
    if (WebInspector.panels.scripts._debuggerEnabled) {
        callback();
        return;
    }

    frontend_addSniffer(WebInspector, "debuggerWasEnabled", callback);
    WebInspector.panels.scripts._toggleDebugging(false);
}

function frontend_checkExecutionLineWhenSourceFrameIsLoaded(testController, callback)
{
    var sourceFrame = WebInspector.currentPanel.visibleView.sourceFrame;
    if (sourceFrame._loaded) {
        testController.results.push("Execution line number: " + sourceFrame.executionLine);
        testController.results.push("Execution line text: " + sourceFrame._textModel.line(sourceFrame.executionLine - 1));
        callback();
    } else {
        setTimeout(function() {
            frontend_checkExecutionLineWhenSourceFrameIsLoaded(testController, callback);
        }, 100);
    }
}

</script>
</head>

<body onload="onload()">
<p>
Tests that debugger will stop on "debugger" statement in a function that was added
to the inspected page via evaluation in Web Inspector console.<a href="https://bugs.webkit.org/show_bug.cgi?id=41058">Bug 41058</a>
</p>

</body>
</html>
